Refactor MedicationsHelper and create MedicationsFormHelper
